chainlit_draft/requirements.txt
chainlit==2.5.5
openai==1.86.0
python-dotenv==1.0.1